Make utility class able to take overrides to support reusing existing billing profiles

+++ datarepo-clienttests/src/main/java/scripts/testscripts/baseclasses/SimpleDataset.java
@@ -7,12 +7,13 @@ import bio.terra.datarepo.model.BillingProfileModel;
 import bio.terra.datarepo.model.DatasetSummaryModel;
 import bio.terra.datarepo.model.DeleteResponseModel;
 import bio.terra.datarepo.model.JobModel;
-import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import runner.config.TestUserSpecification;
 import scripts.utils.DataRepoUtils;
 import scripts.utils.SAMUtils;
+
+import java.util.List;
 
 public class SimpleDataset extends runner.TestScript {
   private static final Logger logger = LoggerFactory.getLogger(SimpleDataset.class);
@@ -22,6 +23,11 @@ public class SimpleDataset extends runner.TestScript {
   protected DatasetSummaryModel datasetSummaryModel;
 
   public void setup(List<TestUserSpecification> testUsers) throws Exception {
+    setup(testUsers, null);
+  }
+
+  public void setup(List<TestUserSpecification> testUsers, BillingProfileModel billingProfile)
+      throws Exception {
     // pick the a user that is a Data Repo steward to be the dataset creator
     datasetCreator = SAMUtils.findTestUserThatIsDataRepoSteward(testUsers, server);
     if (datasetCreator == null) {
@@ -35,10 +41,15 @@ public class SimpleDataset extends runner.TestScript {
     RepositoryApi repositoryApi = new RepositoryApi(datasetCreatorClient);
 
     // create a new profile
-    billingProfileModel =
-        DataRepoUtils.createProfile(
-            resourcesApi, repositoryApi, billingAccount, "profile-simple", true);
-    logger.info("Successfully created profile: {}", billingProfileModel.getProfileName());
+    if (billingProfile == null) {
+      billingProfileModel =
+          DataRepoUtils.createProfile(
+              resourcesApi, repositoryApi, billingAccount, "profile-simple", true);
+      logger.info("Successfully created profile: {}", billingProfileModel.getProfileName());
+    } else {
+      billingProfileModel = billingProfile;
+      logger.info("Using existing profile: {}", billingProfileModel.getProfileName());
+    }
 
     // make the create dataset request and wait for the job to finish
     JobModel createDatasetJobResponse =
@@ -53,6 +64,11 @@ public class SimpleDataset extends runner.TestScript {
   }
 
   public void cleanup(List<TestUserSpecification> testUsers) throws Exception {
+    cleanup(testUsers, true);
+  }
+
+  public void cleanup(List<TestUserSpecification> testUsers, boolean deleteProfile)
+      throws Exception {
     // get the ApiClient for the dataset creator
     ApiClient datasetCreatorClient = DataRepoUtils.getClientForTestUser(datasetCreator, server);
     ResourcesApi resourcesApi = new ResourcesApi(datasetCreatorClient);
@@ -67,7 +83,11 @@ public class SimpleDataset extends runner.TestScript {
     logger.info("Successfully deleted dataset: {}", datasetSummaryModel.getName());
 
     // delete the profile
-    resourcesApi.deleteProfile(billingProfileModel.getId());
-    logger.info("Successfully deleted profile: {}", billingProfileModel.getProfileName());
+    if (deleteProfile) {
+      resourcesApi.deleteProfile(billingProfileModel.getId());
+      logger.info("Successfully deleted profile: {}", billingProfileModel.getProfileName());
+    } else {
+      logger.info("Skipping profile delete because test is using a shared profile");
+    }
   }
 }
